+++ nautilus/NautilusDefaults.cpp
@@ -7,9 +7,9 @@ namespace nautilus {
 
     namespace defaults {
 
-        const double                        CAMERA_SPEED                            = 2.5;
-        const double                        CAMERA_SENS                             = 0.1;
-        const double                        CAMERA_FOV                              = 105.0;
+        const float                         CAMERA_SPEED                            = 2.5f;
+        const float                         CAMERA_SENS                             = 0.1f;
+        const float                         CAMERA_FOV                              = 105.0f;
 
         const glm::vec3                     WORLD_ORIGIN                            = glm::vec3(0.0f);
         const glm::vec3                     WORLD_UP_GL                             = glm::vec3(0.0f, 1.0f, 0.0f);
